changed quant_config

+++ gemma_270m/train.py
@@ -80,7 +80,7 @@ def main():
     # Load model with quantization
     quantization_config = BitsAndBytesConfig(
         load_in_4bit=True,
-        bnb_4bit_compute_dtype=torch.bfloat16,
+        bnb_4bit_compute_dtype=torch.float16,
         bnb_4bit_quant_type="nf4",
     )
